move calculatePosMatrix from painter to transform
