+++ js/render/draw_background.js
@@ -91,7 +91,7 @@ function drawBackground(painter, source, layer) {
             gl.uniform2fv(program.u_offset_b, [offsetBx, offsetBy]);
         }
 
-        painter.setPosMatrix(painter.calculatePosMatrix(coord));
+        painter.setPosMatrix(painter.transform.calculatePosMatrix(coord));
         gl.drawArrays(gl.TRIANGLE_STRIP, 0, painter.tileExtentBuffer.itemCount);
     }
